+++ ChangeLog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Documentation) XML documentation comments on many types
 - (Documentation) Initial setup for DocFX generator
 ### Changed
+- glTF export menu entries moved from `File -> Export` to `Assets -> Export glTF`
 - (Documentation) Split up monolithic docs into multiple markdown files
 - (Documentation) Changelog links to code are now `xref` (for DocFX)
 ### Fixed